-- change writeLock to readLock in ibis::part::buildIndexes and ibis::part::purgeIndexFiles to avoid dead lock in index constructions (a problem reported by Chavdar Kopoev)
git-svn-id: https://codeforge.lbl.gov/anonscm/fastbit/trunk@396 9ec3b2dc-5744-0410-9993-b8723341f1db

+++ examples/ibis.cpp
@@ -4026,7 +4026,7 @@ static void doAppend(const char* dir) {
 	if (ibis::gVerbose > 3 ||
 	    (ibis::gVerbose >= 0 && testing > 0)) {// self test after append
 	    int nth = static_cast<int>(ibis::gVerbose < 20
-				       ? 1+sqrt(ibis::gVerbose)
+				       ? 1+sqrt((double)ibis::gVerbose)
 				       : 3+log((double)ibis::gVerbose));
 	    tbl->buildIndexes();
 	    ierr = tbl->selfTest(nth);
